requirements.txt
aiobotocore==1.4.2
aiohttp==3.8.0
aioitertools==0.8.0
aiosignal==1.2.0
asn1crypto==1.4.0
async-timeout==4.0.0
attrs==21.2.0
attrs-mate==1.0.0
awswrangler==2.12.1
beautifulsoup4==4.10.0
blessed==1.17.6
boto3==1.17.106
botocore==1.20.106
certifi==2021.10.8
chalice==1.26.2
charset-normalizer==2.0.7
click==8.0.3
cramjam==2.4.0
decorator==5.1.0
et-xmlfile==1.1.0
fastparquet==0.7.1
frozenlist==1.2.0
fsspec==2021.11.0
idna==3.3
inquirer==2.7.0
jmespath==0.10.0
jsonpath-ng==1.5.3
lxml==4.6.4
multidict==5.2.0
mypy-extensions==0.4.3
numpy==1.21.4
openpyxl==3.0.9
opensearch-py==1.0.0
pathlib-mate==1.0.1
packaging==21.2
pandas==1.3.4
pg8000==1.21.3
ply==3.11
progressbar2==3.55.0
pyarrow==5.0.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-editor==1.0.4
python-utils==2.5.6
pytz==2021.3
PyYAML==6.0
pynamodb==5.1.0
readchar==2.0.1
redshift-connector==2.0.889
requests==2.26.0
requests-aws4auth==1.1.1
s3fs==2021.11.0
s3transfer==0.4.2
scramp==1.4.1
six==1.16.0
soupsieve==2.3
smart-open==5.2.1
thrift==0.15.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
wrapt==1.13.3
yarl==1.7.2
